+++ nvml-wrapper/src/lib.rs
@@ -3,10 +3,8 @@ A safe and ergonomic Rust wrapper for the [NVIDIA Management Library][nvml] (NVM
 a C-based programmatic interface for monitoring and managing various states within
 NVIDIA GPUs.
 
-```
+```rust
 use nvml_wrapper::Nvml;
-# use nvml_wrapper::error::*;
-# fn test() -> Result<(), NvmlError> {
 
 let nvml = Nvml::init()?;
 // Get the first `Device` (GPU) in the system
@@ -19,8 +17,6 @@ let encoder_util = device.encoder_utilization()?; // Currently 0 on my system; N
 let memory_info = device.memory_info()?; // Currently 1.63/6.37 GB used on my system
 
 // ... and there's a whole lot more you can do. Most everything in NVML is wrapped and ready to go
-# Ok(())
-# }
 ```
 
 NVML is intended to be a platform for building 3rd-party applications, and is
@@ -55,6 +51,25 @@ This wrapper is being developed against and currently supports NVML version
 11. Each new version of NVML is guaranteed to be backwards-compatible according
 to NVIDIA, so this wrapper should continue to work without issue regardless of
 NVML version bumps.
+
+### Legacy Functions
+
+Sometimes there will be function-level API version bumps in new NVML releases.
+For example:
+
+```text
+nvmlDeviceGetComputeRunningProcesses
+nvmlDeviceGetComputeRunningProcesses_v2
+nvmlDeviceGetComputeRunningProcesses_v3
+```
+
+The older versions of the functions will generally continue to work with the
+newer NVML releases; however, the newer function versions will not work with
+older NVML installs.
+
+By default this wrapper only provides access to the newest function versions.
+Enable the `legacy-functions` feature if you require the ability to call older
+functions.
 
 ## MSRV